Refactor media asset workflows
