+++ apps/backend/src/mediaAssets/uploadSessions/index.test.ts
@@ -1,17 +1,17 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 import type pg from "pg";
-import type { DatabaseExecutor, SqlValue } from "../database";
-import { HttpError } from "../shared/errors";
+import type { DatabaseExecutor, SqlValue } from "../../database";
+import { HttpError } from "../../shared/errors";
 import {
   beginMediaAssetUploadSessionCompletionInExecutor,
   recoverMediaAssetUploadSessionCompletionInExecutor,
-} from "./uploadSessions";
+} from "./index";
 import {
   buildMediaBlobStorageKey,
   buildMediaMultipartUploadStagingStorageKey,
-} from "./storageKeys";
-import type { MediaAssetUploadSessionRow } from "./types";
+} from "../storageKeys";
+import type { MediaAssetUploadSessionRow } from "../types";
 
 const testWorkspaceId = "11111111-1111-4111-8111-111111111111";
 const testMediaAssetId = "22222222-2222-4222-8222-222222222222";
